lsusd: allow clients to abort this suspend cycle.
When an event happens in the kernel, it increments the
wakeup_count which aborts the current suspend cycle.
We need something similar for user-space.

So use the atime of the 'disabled' file.  Any process that
triggers an event for another process to read can simply
read from this file and thus abort the current suspend (if there
is one).

Signed-off-by: NeilBrown <neilb@suse.de>