wakealarmd: move timer event from connection to global state.
Rather than having a separate timer event for each active
connection, we now have a single timer event for all connections.

As we keep a sorted list of connections, this is just as easy to work
with.

And it allows us to re-establish the timer after a suspend/resume.
This is needed as the timer might be made more slack by suspend.

Signed-off-by: NeilBrown <neilb@suse.de>