cycle_watchers: don't write the 'all done' signal until after the rename.
When a suspend cycle is finished we move the new 'watching-next' into
place and we write to the old one to make sure any process watching
will notice something has happened.

It is best to write to the file *After* the rename else a process that
notices the file change might make wrong assumptions about the
then-current 'watching' file.

i.e. close a small and probably unimportant race.

Signed-off-by: NeilBrown <neilb@suse.de>